RareList - big update, many fixes
Modify expansion buttons.
Add functions to toggle zones by expansions, use them in expansion buttons.
Some tricky logics to sort zone names, based on which expansion toggled.
Hide scrollbars when it's not needed.
Major thing is to finally fix "CLASSIC", i think i accidentaly put russian C in the end of it so it was looking CLASSIC, but instead there was russian С in the end, which wasted me today at least 6 hours. In fact i fixed it in previous commit, but not all of it.