WowHeadLink, Editbox
WowHeadLink - improve search engine, add #npcs and #zones.
Editbox:
Make the edit box be perfectly aligned to the top of the page 1.
Improve OnChar scripts for editbox, now if user WASD, aka tries to move, the editbox will be hidden.
Clamp it to screen, to avoid user not seeing with panel too high.

Signed-off-by: Sattva <Sattvagit@gmail.com>